Move particle system parsing into more appropriate classes
